refactor: as-is 를 제거해도 정상적으로 동작하도록 수정
