--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ForwardController.java
@@ -1,20 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.asis;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.Objects;
-
-public class ForwardController implements Controller {
-
-    private final String path;
-
-    public ForwardController(final String path) {
-        this.path = Objects.requireNonNull(path);
-    }
-
-    @Override
-    public ModelAndView execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return new ModelAndView(path);
-    }
-}
