+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerMapping.java
@@ -1,6 +1,5 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface HandlerMapping {
